Made crashlytics_path docs less ambiguous.
I encountered an error when I tried to use Crashlytics Beta because I took ```crashlytics_path: './path', # path to your 'Crashlytics.framework'``` to mean the path to the folder where the framework exists, NOT the path to the framework itself. I worry that others may encounter the same confusion, so I think this small change might be less ambiguous.