Improve support for xcpretty report options.
This commit addresses the following issue:
Although a user can set multiple xcpretty reports with the ":report_formats" argument, it is unable to accept multiple values for ":report_path". This means a user cannot only define the output path for one of the reports even if they specify more than one.

This commit defines a new parameter ":reports", which takes an array of hashes. The keys of the has map exactly to the parameters taken by xcpretty. This enables the user to specify all options for each report they want to generate.