Create install directory, if it doesn't exist
When using sigh to install provisioning profiles on a system that is
fresly installed with Xcode, it is entirely possible that the
provisioning profile directory has not been created yet.

If we attempt to simply copy the downloaded profile into that directory
when it does not exist, will cause a failure. Instead we should see if
the directory exists before copying into it. If it does not exist,
we create it, and continue the copy operation.