re-write how the attr_mapping works:
Before: we would define mapped get/set methods on object initialization. This would overwrite any methods of the same name.

After: we extend the class using Module#include.
This allows us to redefine the methods after they have been created by `attr_mapping`. To call the original method, call `super` in the redefinition as if it we a subclass.