Zip all dSYM files together
When building an application with many frameworks, each framework
will produce it's own dSYM file. These need to be included in the
dSYM zip file to allow for symbolication of crashes in these
frameworks.

Really we just need the archive path, but this will allow for
easy backward compatibility with existing code.