Trim whitespace & newlines in locale string
The issue mentioned in #343 is most likely caused due to a new line in the "/tmp/language.txt" file. At least this is what happened to me when manually editing the file & trying to run tests in Xcode. A newline at the end of locale string caused the AppleLocale parameter to be split into two lines which prevented the simulator to apply the correct locale to the app and falling back to "en_US". Trimming new lines fixes this issue.