Reimplement overview parsing in light of ITC API changes (#228)
For application, :versions in summary field is now replaced by :versionSets. This forces us
to reimplement the :format field creatively to the best of our capabitity, by reusing the heuristic
we used under client.app_version

We obviously also update the data used in fixtures to match the new formats.