Reverse the default value of :shallow_clone to be false
There have been a number of reports that doing a shallow clone
in git leaves match unable to push changes back to the remote
repository. This manifests as a fatal error in match saying,
"protocol error: expected old/new/ref, got 'shallow xxxxxx'".

Because shallow cloning is an optimization that is causing
problems for some users, I think it is best if we reverse
the current default value for this option to be off, to fix
possible problems, but allow opt-in for the optimizations.

This should be backwards-compatible in that folks who are
not specifying a value for shallow_clone will stop getting
shallow clones, but continue to work.