1) use_webhook_configured_username_and_icon: default value is false 2) username: default value is fastlane 3) icon_url: default value is https://s3-eu-west-1.amazonaws.com/fastlane.tools/fastlane.png
The current implementation always overrides the username and icon_url
settings.

If use_webhook_configured_username_and_icon is set to true, then the
action will fall back to the values set in the webook's settings in
Slack. Otherwise a username and/or icon_url can be passed in to use in
place of the default values.