Change order of screenshot name
Current implementation has the screenshot name as
<timestamp>_<user-defined-name>.png - this means that the screenshots will be
ordered in time order when uploaded to Google Play. On top of this, the order
will be decided by the naming of the tests as the tests are run in alphabetical
order.

This change adds the timestamp as a suffix instead, which means the user can
choose the order of the screenshots by using the screenshot name setting.