Update Bamboo.md
I've been using fastlane carthage and bamboo for a few months now and one of the issue's we've run into is that build times can start to bog down with using carthage.  What we've done is separate all our build plans into multiple plans as described above.  This has saved us countless hours (like 15 mins per build) and I wanted to share the knowledge with the general public.  I know it might be a "little" specific so let me know what you think - but I figured this was better served in your repo than carthage because fastlane pairs so well with CI.