Make body of commit optional in jenkins changelog generation
A list of commit summaries is a much better changelog than a list of
commit summaries + their bodies. There are a few reasons for this. For
one, most obvious, a changelog is ideally a summary of changes. Another
is there might be restrictions on the character length of your
changelog. Crashlytics for example has a limit of ~16k characters, so if
everyone wrote really detailed bodies in their commits, the build would
fail after a few commits. Also, it is a lot easier to read and
understand.