Make deliver setup honor the skip_screenshots flag (#4919)
With this commit calling

  deliver init --skip_screenshots

will only download metadata and skip downloading of screenshots. This is
useful for cases where the user is only interested in manipulating and
uploading the metadata as it spares them the lengthy screenshot download.
Before, the only option to skip screenshots during initialization was
to CTRL+C the process after the metadata was downloaded.