Make CSV header optional in import_testers (#4119)
`pilot import` always skips the first row of a CSV file because it
expects the format

    header1,header2,header3
    column1,column2,column3

whereas the iTunesConnect bulk import function expects a CSV with
format

    column1,column2,column3

This means that if `pilot import` is used with an iTC CSV, it will
skip the first tester in the file.

This patch changes the skipping logic. Instead of skipping the first
row, it skips any row that doesn't have a valid email on the email
column.

By making the header effectively optional, `pilot import` can
successfully process both an iTC CSV and the resulting CSV of `pilot
export`.