Improve junit template reference style
The previous style would return './' for the starting location
if the fastlane gem was not installed. This does not work when
running fastlane relative to another project directory.

The style it is changed to is what I believe to me a more
common Ruby style for referring to needed files at a relative
path.