[scan] Handle single scheme Xcode projects
Single scheme projects are explained here: http://promisekit.org/news/2016/08/Multiplatform-Single-Scheme-Xcode-Projects/

With a single scheme project, it is possible that both `Scan.project.ios?` and `Scan.project.tvos?` return false even though iOS and tvOS are supported platforms for the project. So instead of detecting simulators if the project supposedly support iOS or tvOS, simulators are detected when the --device or --devices option is passed to scan.