Allow xcargs to accept Hash & Array, auto-converted to a shell-escaped String (#8051)
`ConfigItems` can now accept their `type` to be the special `:shell_string` symbol, meaning that they expect a `String`, but if given a `Hash` or an `Array`, those will be auto-converted to a shell-escaped String for you.

This is especially used for the parameter `xcargs` of actions like `ipa`, `gym`, `scan` & `snapshot`, helping clarity a lot as they now can accept e.g. Hashes instead of having to escape and compose long strings manually.