Catch and print exceptions occuring in error block
This has been annoying, and really hard to debug: when the error block raised an exception, it basically replaced the actual reason of error. The error block is usually used to notify about the failure, but the user is interested in the *actual* error message.

This PR also adds a test covering exactly that