Fix sigh download_all options regression
This was caused by the refactoring done in #8129.
I incorrectly thought that renew was the only
command that used the sigh tool options. While I
added tests to confirm that renew and resign did
not conflict anymore, I did not add tests for the
other commands.

This now correctly applies the sigh tool options
for download_all and repair, which also need them