Add "enterprise" reference to the match documentation
`match` can deal with certificates and provisioning profiles for
development, adhoc, appstore and inhouse builds.
In the `README.md` as well as the generated `MatchFile` no mention are
done about the `enterprise` type.

The goal of this request is to facilitate `match`'s usage for enterprise
users by quickly showing them that this option exists and how to use it.