Setup fastlane train for fastlane deployment
This allows us to see when someone in the team has a deployment process running (which takes about 10 minutes), so that no one else merges anything into the master branch. If something is merged while something is deployed, it might not show up in the changelog.

If you feel like it pollutes the Fastfile too much, we can also remove the PR, or simplify it to be less steps