[match] Add new `clone_branch_directly` option to clone faster
When an app developer used their main repo, and selected a different branch for match to store the certificates, it becomes really slow very fast if the `master` branch has bigger resources. In my case it took many minutes already.

This PR adds a parameter that tells match that it should just clone this one branch. Also we'll now show a message that tells users about this new option