Show warning when no value is provided for option in config files (#9108)
Assuming a `Gymfile` with the content:
```
scheme "Release"
clean
```

The `clean` will be completely ignored, even though the user would assume it might set a boolean flag to true. Currently we just silently do nothing. We should show a warning that no value was provided and tell the user how to set one.

This PR also adds a test covering this change