Add additional guard clause to pricing_info
If an in-app purchase is not yet "cleared for sale", Apple's API does
not return any pricing information. So for these IAPs,
"pricing_intervals" on IAPDetail returns nothing (an empty array []).
Therfore it's also not possible to retrieve any actual pricing
information with "pricing_info".
This adds an additional guard clause to "pricing_info" checking for the
"cleared_for_sale" flag in order to prevent unnecessary API calls and
computation. To prevent unnecessary errors, the private API of IAPDetail
is now also more defensive when checking whether an IAP uses world wide
pricing or not.