Update to the latest version of the Google API Client
Version 0.12.0 of the Google API Client gem has breaking API changes. Unfortunately, the previous gemspec requirement (~> 0.11) was slightly too permissive, and allowed the new version to be pulled in accidentally.

While I'm in here, it seemed like a good idea to update to 0.12.0 and fix the breakages. But the simplest possible fix would just be to lock the gem to 0.11.x