[spaceship] Fix reporting successful submission even it fails due to problems with Ad ID information
iTunes Connect still returns a success status code even the submission
was failed because of Ad ID info.  This checks for any section error
keys in returned adIdInfo.

This resolves #8260, where `deliver` says it successfully submitted
application for review when it really didn't. It should return an error
with appropriate info.