[Gemspec] Restrict public_suffix to version 2.0.x
public_suffix a gem which we require transitively through
google-api-client -> adressable -> public_suffix recently released
version 3.x, which removes support for ruby 2.0.0.

Because addressable requires <= 4.0.0, >= 2.0.2, it is safe to
pin this to the 2.x versions for now, to fix installation of fastlane on
macOS system ruby.

If preferred, we could also make this conditional based on the ruby
version, but I am not a fan of complicating dependency graphs.