Fix issue with the order of where values are taken from
Fixes https://github.com/fastlane/fastlane/issues/10640

Documented here https://docs.fastlane.tools/advanced/#priorities-of-parameters-and-options

This PR doesn't change or add any tests yet

First, I want to have a discussion, as this change probably causes side effects for some setups: It seems like we haven't changed the behavior since many months (or years?), so I'd assume some setups already make "use" of that behavior.

Assuming that's true, and we haven't changed anything in that code for the last year, I'd propose to change the documentation, instead of applying this PR, as both solutions are fine, as long as we're consistent and have the proper documentation for it.