Allow snapshots on older iOS versions when no specified/latest found (see #10211)
This allows generating snapshots on older ios versions for a simulator
if the specified or latest ios version is not found for that particular
device. The work was already done by the find_device method but wasn't
being used by the destination method.
This fixes iPhone 4s screenshots not being generated on XCode 9 (#10211)
Fixes #10211