Enable auto complete when using bundler
This enables typing "bundle exec fastlane lane_nam" and pressing
tab to auto complete with "lane_name".

The trade-off is that it also works when typing
"bundle exec lane_nam" or even "bundle lane_nam", but I think it's
an acceptable trade-off, as the chances of something thinking it
can work this way seems low.