Fix FastlaneCore::Helper.xcode_path for non-Mac environments
`FastlaneCore::Helper.xcode_path` finds out if it is inside a test by checking for if `SpecHelper` is defined:
https://github.com/fastlane/fastlane/blob/5019acb53fc2adf050b4227c3e3216e365ba36ec/fastlane_core/lib/fastlane_core/helper.rb#L23-L25
If the module is only loaded after the check is first executed, it internally never realizes it is in a test and return the wrong value on non-Mac environments. Only on subsequent calls does it return the correct value.

Moving the SpecHelper module code before the first call to `FastlaneCore::Helper.xcode_path` fixes the problem.