Improve the command timeout behavior by progressively increasing the timeout value on each iteration. This stands a greater chance of succeeding than by simply repeating the command with the same timeout several times.
This commit also reduces the default timeout values from 10 to 3. The choice of 3 is somewhat arbitrary but in practice it ends up causing a maximum timeout total of 45 seconds vs. the old maximum timeout of 40 seconds. Yet while the old approach only ever waits as long as 10 seconds for a command to complete, the new approach will wait as long as 24 seconds, greatly increasing the odds of a successful result.

I also updated the messages that are displayed when timeouts occur, to reflect the number of tries that are being made and the number of seconds being waited for each try. A typical example of a final try message is here:

Command timed out after 12 seconds on try 3 of 4, trying again with 24 second timeout...