Fix deliver(overwrite_screenshots: true) failing with iMessage screenshots
We were hitting the same error as #10032. After some debugging, it seems like when #6981 was done (in #7165/#7162), it always passed `false` for the `is_messages` parameter of the `upload_screenshot` method (https://github.com/fastlane/fastlane/pull/7162/files#diff-1d448c7e200ec0facd4ffd5c23cd3fdeR21)

This is not covered in the tests, if you don't mind pointing at how/where I should best do that, I would be happy to include that in this PR as well!