Update docs on gym's :export_options action parameter
The :export_options cannot be a path to a plist anymore when running gym as a Fastlane action.
It was OK before when this config item was consumed only to prepare a file for xcodebuild
where it would be read from disk in case it was not a hash. However now it is used in
DetectValues.set_additional_default_values which does not expect it to be a non-hash.