map errors to displayable_errors
This gets rid of errors_to_show but we are still getting errors that are essentially double printed.
What's happening is the error line(s) from transporter are getting put into @errors, and then UI.crash! is being rescued and and that is being put into @errors as well.

On top of this there are a few spots that either check if @errors is > 0 and prints the errors/tells you there were errors and also places that check the exit status and then prints the errors/tells you there were errors