Add support for release notes from bundles
For projects that use AAB instead of APK, the release notes functionality of fastlane wasn't working. This is because the `upload_changelogs` functionality specifically looks for APK version codes - which obviously don't include AABs. This PR fixes this by adding support for a new client library function to list the AAB versions and invoking this in the `uploader`.

Another approach would be to change the `apk_version_codes` client library call to also include bundles. This felt like it would be more disruptive for other use cases.