[get_version_number] look into xcconfig for plist
When `get_version_number` uses Xcodeproj's `target.resolved_build_setting` to look for the `INFOPLIST_FILE` build setting, it will crash if the value can't be found, as can happen if the value is defined in xcconfig files. I resolved this by adding a test to make sure we do find a value, and by also checking inside xcconfig files.