[precheck] Allow mentioning Google Drive (#13754)
### Checklist

- [x] I've run `bundle exec rspec` from the root directory to see all new and existing tests pass
- [x] I've followed the _fastlane_ code style and run `bundle exec rubocop -a` to ensure the code style is valid
- [x] I've read the [Contribution Guidelines](https://github.com/fastlane/fastlane/blob/master/CONTRIBUTING.md)
- [x] I've updated the documentation if necessary.

### Motivation and Context

This fixes a false positive in `precheck`. It is reasonable that apps might integrate with Google Drive and users would want to know about this from the App Store description. Our app PDF Viewer has been approving mentioning Google Drive.

### Description

I added an additional entry to the whitelist for mentioning other platforms. Google is on the blacklist because it is a competitor to Apple, then Google Analytics and Google Drive are whitelisted because iOS app may integrate with these services.

I added tests for `Precheck::OtherPlatformsRule`, just covering a one allowed case and one not allowed case. I don’t really know Ruby or any of the rig here so I copied and modified `curse_words_rule_spec.rb`.