[zip] Handle symbolic links (#13836)
With symlinks option it is now possible to create a zip file that stores symbolic links to the archive

<!-- Thanks for contributing to _fastlane_! Before you submit your pull request, please make sure to check the following boxes by putting an x in the [ ] (don't: [x ], [ x], do: [x]) -->

### Checklist
- [x] I've run `bundle exec rspec` from the root directory to see all new and existing tests pass
- [x] I've followed the _fastlane_ code style and run `bundle exec rubocop -a` to ensure the code style is valid
- [x] I've read the [Contribution Guidelines](https://github.com/fastlane/fastlane/blob/master/CONTRIBUTING.md)
- [x] I've updated the documentation if necessary.

### Motivation and Context
Fastlane comes with a Zip action that enable users to zip files and folders. But it does not handle symbolic links as links, but instead it include all linked files. This is not wanted when the folder to zip already contains the linked files.

### Description
I have improved the already existing Zip action to control whether symbloic links should be treated as symbolic links or not. 
By default the behaviour of the action hasn't changed. By setting symlinks: true the user will get a zip file that can contain symbolic links.

I tested this on my build system by creating a copy of the zip action and included it in my repo. It works as expected :)