resign.sh - Fix bundle ID update in patched entitlements
When the new bundle ID is a prefix of the old bundle ID, and the patched
entitlements file contains the correct ID already, the code old ended up
duplicate the suffix in the entitlements.

e.g. old ID: com.example.foo
     new ID: com.example.fooBar

Would result in a bundle ID of com.example.fooBarBar in the entitlements
file.

Fixes #9550