Fix support for specifying emulator serial (#13946)
* Fix support for specifying emulator serial

The `strip!` function returns nil if the string was not altered. (https://ruby-doc.org/core-2.2.0/String.html#method-i-strip-21).  The adb action now fails whenever a serial _is_ specified with `no implicit conversion of nil into String` on line 42 (the Actions.sh call).

Since command is being set with the return value, `strip` should be used instead which will always return the new value, regardless of whether it was changed.

* Update adb_spec.rb

Add test cases to validate the expected behavior when a serial is passed.