gemspec: Raise upper limit on Bundler version to 3.0.0
I can't really see any issues resulting from this.  Upgrading it didn't
seem to clearly break anything when I ran tests, so this seems ready to
at least try and get CI to run on.

Note that 2.0.1 will not install on Ruby < 2.3 or RubyGems <= 2.5.0, so
this only allows those users with updated versions of these to upgrade.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>
Tested-by: Kristofer Rye <kristofer.rye@gmail.com>