[automatic_code_signing] create TargetAttributes for targets missing them (#12782)
<!-- Thanks for contributing to _fastlane_! Before you submit your pull request, please make sure to check the following boxes by putting an x in the [ ] (don't: [x ], [ x], do: [x]) -->

### Checklist
- [x] I've run `bundle exec rspec` from the root directory to see all new and existing tests pass
- [x] I've followed the _fastlane_ code style and run `bundle exec rubocop -a` to ensure the code style is valid
- [x] I've read the [Contribution Guidelines](https://github.com/fastlane/fastlane/blob/master/CONTRIBUTING.md)
- [x] I've updated the documentation if necessary.

### Motivation and Context
<!-- Why is this change required? What problem does it solve? -->
I have a project which had a `target` but did not have a `TargetAttributes` entry for that `target`. This was a new Cordova project.

<!-- Please describe in detail how you tested your changes. -->
I tested by creating a new project which exhibited the same issue as this problem solved. I used this code branch instead and was able to use the `automatic_code_signing` action on the target which I was previously having an issue with.

### Description
<!-- Describe your changes in detail -->
This change ensures a `TargetAttributes` entry exists for any `target`. Before looping through all `TargetAttributes` entries, this goes through all `Targets` and ensures an entry exists in `TargetAttributes`


Co-authored-by: Jan Piotrowski <piotrowski+github@gmail.com>