skipping the search for the crashlytics API token if parameter gsp_path is set
Since the search for the API token will search recursively in all Info.plist files,
including all projects in the Pods directory, the search might produce confusing
console output like 'invalid byte sequence in UTF-8' if a file cannot be read.
When the parameter gsp_path is set, there is no need to search for the API token
and we can skip the search, like it is already done in function find_gsp_path
if parameter api_token is set.