Add a read timeout option for artifactory
The Artifactory gem supports a configurable option for the read timeout.  The Fastlane action can be extended to pass the argument through to the client.

The [default](https://github.com/chef/artifactory-client/blob/master/lib/artifactory/defaults.rb#L149) read timeout for the client is 120 so I see no reason to set a different default.