Fix Supply Error When No Tracks Are Returned
I fixed an error in Supply's Client.tracks method. When there are no
active releases for an application in the Google Play console, the call
to the Edit.tracks:list Google Play API succeeds, but the JSON response
is not including the tracks array. This results in the all_tracks
variable being nil and the call to all_tracks.select fails with a nil
reference error. I added a line that will default the all_tracks
variable to an empty array if all_tracks is nil which fixes the error.