Add 5 code-examples to documentation, update details, authors
It took me some time and trial & error to figure out how to use this action to modify the values inside a plist-file, whose functionality was (imho) only sparsely documented.
I added 5 new examples for the basic functionality of this action, in addition to the one (more complicated and highly specific) example already present, to improve understandability and readability in the future. I also slightly adapted the details in order to differentiate this action from the `update_info_plist`action.