Fixing empty track use case
The Google Play API returns a different error message body now, and the value of "e.to_s" is now something like this: "notFound: Track not found: alpha". I don't know if the "trackEmpty" case still applies, but I left it there in case.
!!! Please verify and test this before merging !!!
I have not tested this myself and this is just how I assume it should be fixed, maybe it's something completely different.

This is meant to fix a bug that is now happening: the "google_play_track_version_codes" lane will error out if querying for the version codes list of an empty track.